requirements.txt

requests
awscli
awslambdaric



boto3
Pillow
numpy==1.26.4
torch==2.1.0
torchvision==0.16.0
realesrgan
basicsr
facexlib
gfpgan
wheel
addict
future
lmdb
pyyaml
scikit-image
tb-nightly
yapf
filterpy
numba
llvmlite
matplotlib
contourpy
cycler
fonttools
kiwisolver
pyparsing
python-dateutil
importlib-resources
six
opencv-python==4.9.0.80